refactor: split cache manager tests
